perf: 生成速度を向上

+++ packages/icons-webfont/.build/build-outline.mjs
@@ -5,7 +5,7 @@ import { resolve, basename } from 'path'
 import crypto from 'crypto'
 import { glob } from 'glob'
 import { execSync } from 'child_process'
-import Queue from 'promise-queue'
+import { execa } from 'execa'
 
 const DIR = getPackageDir('icons-webfont')
 
@@ -25,11 +25,10 @@ const buildOutline = async () => {
     const stroke = strokes[strokeName]
 
     await asyncForEach(Object.entries(icons), async ([type, icons]) => {
-      fs.mkdirSync(resolve(DIR, `icons-outlined/${strokeName}/${type}`), { recursive: true })
+      fs.mkdirSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/new`), { recursive: true })
       filesList[type] = []
-      const queue = new Queue(32)
 
-      const promises = icons.map(({ name, content, unicode }) => queue.add(async function () {
+      await asyncForEach(icons, async function ({ name, unicode, content }) {
         console.log(type, name);
 
         if (compileOptions.includeIcons.length === 0 || compileOptions.includeIcons.indexOf(name) >= 0) {
@@ -66,9 +65,14 @@ const buildOutline = async () => {
 
               // Check hash
               if (crypto.createHash('sha1').update(cachedContent).digest("hex") === cachedHash) {
-                console.log('Cached stroke for:', name, unicode)
+                console.log('Cached completed stroke for:', name, unicode)
                 return true;
               }
+            }
+
+            if (unicode && fs.existsSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/new/${cachedFilename}`))) {
+              console.log('Cached outlined stroke for:', name, unicode)
+              return true;
             }
 
             await outlineStroke(content, {
@@ -80,31 +84,48 @@ const buildOutline = async () => {
               color: 'black'
             }).then(outlined => {
               // Save file
-              fs.writeFileSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/${filename}`), outlined, 'utf-8')
-
-              // Fix outline
-              execSync(`fontforge -quiet -lang=py -script .build/fix-outline.py icons-outlined/${strokeName}/${type}/${filename}`).toString()
-              execSync(`svgo icons-outlined/${strokeName}/${type}/${filename}`).toString()
-
-              // Add hash
-              const fixedFileContent = fs
-                .readFileSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/${filename}`), 'utf-8')
-                .replace(/\n/g, ' ')
-                .trim(),
-                hashString = `<!--!cache:${crypto.createHash('sha1').update(fixedFileContent).digest("hex")}-->`
-
-              // Save file
-              fs.writeFileSync(
-                resolve(DIR, `icons-outlined/${strokeName}/${type}/${filename}`),
-                fixedFileContent + hashString,
-                'utf-8'
-              )
+              fs.writeFileSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/new/${filename}`), outlined, 'utf-8')
+              console.log('Created stroke for:', name, unicode)
             }).catch(error => console.log(error))
           }
         }
-      }))
+      })
 
-      await Promise.all(promises)
+      // Process for new files
+      // Fix outline
+      if (
+        fs.existsSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/new`)) &&
+        (await glob(resolve(DIR, `icons-outlined/${strokeName}/${type}/new/*.svg`))).length > 0
+      ) {
+        await execa('fontforge', ['-quiet', '-lang=py', '-script', '.build/fix-outline.py', resolve(DIR, `icons-outlined/${strokeName}/${type}/new`)], {
+          stdout: process.stdout,
+          stderr: process.stderr,
+        });
+        await execa('pnpm', ['dlx', 'svgo', resolve(DIR, `icons-outlined/${strokeName}/${type}/new`)], {
+          stdout: process.stdout,
+          stderr: process.stderr,
+        });
+
+        // Add hash
+        await asyncForEach((await glob(resolve(DIR, `icons-outlined/${strokeName}/${type}/new/*.svg`))), async (dir) => {
+          const filename = basename(dir)
+          const fixedFileContent = fs
+            .readFileSync(resolve(DIR, `icons-outlined/${strokeName}/${type}/new/${filename}`), 'utf-8')
+            .replace(/\n/g, ' ')
+            .trim();
+          const hashString = `<!--!cache:${crypto.createHash('sha1').update(fixedFileContent).digest("hex")}-->`
+
+          // Save file
+          fs.writeFileSync(
+            resolve(DIR, `icons-outlined/${strokeName}/${type}/${filename}`),
+            fixedFileContent + hashString,
+            'utf-8'
+          )
+        })
+      }
+
+      // Remove new directory
+      execSync(`rm -rf icons-outlined/${strokeName}/${type}/new`)
     })
 
     // Remove old files
